--- app/controllers/teachers/modules_controller.rb
@@ -1,9 +1,0 @@
-module Teachers
-  class LModulesController < BaseController
-    def index
-    end
-
-    def present
-    end
-  end
-end
