+++ app/views/teachers/classrooms/participant_surveys/_participant_survey.html.slim
