+++ app/views/teachers/classrooms/quizzes/index.html.slim
@@ -1,4 +1,4 @@
 - content_for :classroom_show_content do
-  = render @quizzes
+  = render @quizzes, classroom: @classroom
 
 - parent_layout "classroom_show"
