--- app/views/teachers/quizzes/_quiz.html.slim
@@ -1,5 +1,0 @@
-/ Local vars:
-/   quiz
-a.survey--blue-box href="#" target="_blank"
-  h1.survey__title = quiz.name
-  p.survey__description = quiz.description
